Refactor catalog router to use separate modules
